# public/docs/admin/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Error": {
				"description": "Represents an error in the API response.",
				"properties": {
					"code": {
						"type": "string",
						"description": "The error code."
					},
					"message": {
						"type": "string",
						"description": "The error message."
					}
				},
				"required": [
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Warning": {
				"description": "Represents a warning in the API response.",
				"properties": {
					"code": {
						"type": "string",
						"description": "The warning code."
					},
					"message": {
						"type": "string",
						"description": "The warning message."
					}
				},
				"required": [
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LanguagesData": {
				"description": "Represents the data structure for languages.\nExtends List with a generic parameter of LanguageInfo type.",
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/LanguageInfo"
						},
						"type": "array",
						"description": "An array of items."
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetLanguagesRes": {
				"description": "Represents the response payload for retrieving a list of languages.\nExtends the base ApiResponse class with a generic parameter of LanguagesData type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/LanguagesData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CountriesData": {
				"description": "Represents the data structure for countries.\nExtends List with a generic parameter of CountryInfo type.",
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/CountryInfo"
						},
						"type": "array",
						"description": "An array of items."
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetCountriesRes": {
				"description": "Represents the response payload for retrieving a list of countries.\nExtends the base ApiResponse class with a generic parameter of CountriesData type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/CountriesData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"CountryData": {
				"description": "Represents detailed information about a country.",
				"properties": {
					"country": {
						"$ref": "#/components/schemas/CountryInfo",
						"description": "Information about the country."
					},
					"states": {
						"items": {
							"$ref": "#/components/schemas/StateData"
						},
						"type": "array",
						"description": "List of states within the country."
					}
				},
				"required": [
					"country",
					"states"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"country": {
						"iso2": "US",
						"iso3": "USA",
						"name": "United States"
					},
					"states": [
						{
							"state": {
								"id": 1,
								"name": "California"
							},
							"cities": [
								{
									"id": 123,
									"name": "Los Angeles"
								}
							]
						}
					]
				}
			},
			"GetCountryRes": {
				"description": "Represents the response payload for retrieving information about a country.\nExtends the base ApiResponse class with a generic parameter of CountryData type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/CountryData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"GetPermissionsRes": {
				"description": "Represents the response payload for retrieving a list of permissions.\nExtends the base ApiResponse class with a generic parameter of PermissionsData type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/PermissionsData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"GetPermissionRes": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/PermissionData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"AddRoleRes": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AddRoleData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"RoleReq": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"accessControl": {
						"$ref": "#/components/schemas/AccessControlReq"
					}
				},
				"required": [
					"name",
					"accessControl"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"name": "Admin",
					"description": "Adminstrator Role",
					"accessControl": {
						"permissions": [
							{
								"permissionId": "61e7eb4be4a1e5b7c8db7fd1",
								"accessType": "FullControl"
							},
							{
								"permissionId": "61e7eb4be4a1e5b7c8db7fd2",
								"accessType": "ReadOnly"
							}
						],
						"hasAllPermissions": true
					}
				}
			},
...
			"GetRolesRes": {
				"description": "Represents the response payload for retrieving a list of roles.\nExtends the base ApiResponse class with a generic parameter of RolesData type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/RolesData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"GetRoleRes": {
				"description": "Represents detailed information about a role.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/RoleData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRoleRes": {
				"description": "Represents the response payload for updating a role.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Success"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DeleteRoleRes": {
				"description": "Represents the response payload for deleting a rol.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Success"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"AccessToken": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header",
				"scopes": {
					"admin": "Admin"
				}
			},
			"RefreshToken": {
				"type": "apiKey",
				"name": "x-refresh-token",
				"in": "header",
				"scopes": {
					"admin": "Admin"
				}
			}
		}
	},
	"info": {
		"title": "Node.js tsoa MongoDB Boilerplate Admin API",
		"version": "v1",
		"description": "This is a Node.js tsoa MongoDB Boilerplate Admin API version v1. It provides comprehensive documentation and administrative functionalities for managing MongoDB databases. The API is designed to streamline administrative tasks and enhance the overall efficiency of MongoDB operations. The 'public/docs/admin' directory serves as the output location for the API documentation, ensuring easy access for developers and administrators. With specVersion 3, this API adheres to the latest specifications, offering robust performance and compatibility. Whether it's managing user permissions, data schemas, or database configurations, this API simplifies administrative workflows and empowers developers to build scalable MongoDB applications with ease.",
		"license": {
			"name": "Apache 2.0"
		},
		"contact": {
			"name": "Akram B.",
			"email": "belhajsalem.akram@gmail.com"
		}
	},
	"paths": {
...
		"/directories/languages": {
			"get": {
				"operationId": "getLanguages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetLanguagesRes"
								}
							}
						}
					}
				},
				"description": "Get supported languages.",
				"summary": "Get a list of supported languages.",
				"tags": [
					"Multilingual Directories"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
		"/directories/countries": {
			"get": {
				"operationId": "getCountries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCountriesRes"
								}
							}
						}
					}
				},
				"description": "Get supported countries.",
				"summary": "Get a list of supported countries.",
				"tags": [
					"Multilingual Directories"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
		"/directories/countries/{iso3}": {
			"get": {
				"operationId": "getCountry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCountryRes"
								}
							}
						}
					}
				},
				"description": "Get information about a specific country.",
				"summary": "Get information about a specific country by ISO3 code.",
				"tags": [
					"Multilingual Directories"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The ISO3 code of the country to retrieve information about.",
						"in": "path",
						"name": "iso3",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/permissions": {
			"get": {
				"operationId": "getPermissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetPermissionsRes"
								}
							}
						}
					}
				},
				"description": "Get a list of permissions.",
				"summary": "Get a list of permissions based on query parameters.",
				"tags": [
					"Permission Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "The page number for paginated results.\nIt must not be empty, should be an integer, and must be greater than or equal to 1.",
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The number of permissions to be displayed per page.\nIt must not be empty, should be an integer, and must be greater than or equal to 5.",
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The search query for filtering permissions based on name or other criteria (optional).",
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/permissions/{permissionId}": {
			"get": {
				"operationId": "getPermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetPermissionRes"
								}
							}
						}
					}
				},
				"tags": [
					"Permission Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "permissionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/roles": {
			"post": {
				"operationId": "addRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddRoleRes"
								}
							}
						}
					}
				},
				"description": "Adds a new role based on the provided request body.",
				"summary": "Add a new role.",
				"tags": [
					"Role Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The request body containing information about the role.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoleReq",
								"description": "- The request body containing information about the role."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "getRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetRolesRes"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of roles.",
				"summary": "Get a list of roles.",
				"tags": [
					"Role Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "The page number for paginated results.\nIt must not be empty, should be an integer, and must be greater than or equal to 1.",
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The number of roles to be displayed per page.\nIt must not be empty, should be an integer, and must be greater than or equal to 5.",
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The search query for filtering roles based on name or other criteria (optional).",
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/roles/{roleId}": {
			"get": {
				"operationId": "getRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetRoleRes"
								}
							}
						}
					}
				},
				"description": "Retrieves information about a specific role based on the provided role ID.",
				"summary": "Get role details.",
				"tags": [
					"Role Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The unique identifier of the role.",
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "updateRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateRoleRes"
								}
							}
						}
					}
				},
				"description": "Updates information about a specific role based on the provided role ID and request body.",
				"summary": "Update role details.",
				"tags": [
					"Role Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The unique identifier of the role.",
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The request body containing updated information about the role.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoleReq",
								"description": "- The request body containing updated information about the role."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteRoleRes"
								}
							}
						}
					}
				},
				"description": "Deletes a specific role based on the provided role ID.",
				"summary": "Delete a role.",
				"tags": [
					"Role Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The unique identifier of the role to be deleted.",
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/admin/v1"
		}
	],
...
}